# Distributed Systems/event_driven_architecture/java/examples/package-tracker-mono-backend/gateway-service/target/classes/static/openapi.yaml
openapi: 3.0.3
info:
  title: Package Tracker API
  description: |
    Unified API Gateway for the Package Tracker system.
    Exposes Customer, Shipment, Delivery, and Notification services through a single endpoint.
  version: 1.0.0
servers:
  - url: http://localhost:8086
    description: Local development gateway
paths:
...
  /api/v1/orders:
    get:
      tags: [Orders]
      summary: List all orders
      operationId: getAllOrders
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter orders by user ID
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      tags: [Orders]
      summary: Create a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
    Address:
      type: object
      properties:
        street:
          type: string
          example: "221B Baker Street"
        city:
          type: string
          example: "London"
        state:
          type: string
          example: "Greater London"
        zip:
          type: string
          example: "NW16XE"
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: "Priya Sharma"
        email:
          type: string
          format: email
          example: "priya.sharma@example.com"
        phone:
          type: string
          example: "+91-9123456789"
        address:
          $ref: "#/components/schemas/Address"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    OrderStatus:
      type: string
      enum: [PENDING, CREATED, CONFIRMED, PROCESSING, SHIPPED, DELIVERED, CANCELLED]
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: string
          example: "ORD-2026-1001"
        user:
          $ref: "#/components/schemas/User"
        status:
          $ref: "#/components/schemas/OrderStatus"
        totalAmount:
          type: number
          format: double
          example: 2499.99
        shippingAddress:
          $ref: "#/components/schemas/Address"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required: [user, totalAmount]
      properties:
        orderId:
          type: string
          description: Optional custom order ID (auto-generated if omitted)
          example: "ORD-2026-1001"
        user:
          type: object
          required: [id]
          properties:
            id:
              type: integer
              format: int64
              example: 1
        status:
          $ref: "#/components/schemas/OrderStatus"
          description: Optional status (defaults to CREATED)
        totalAmount:
          type: number
          format: double
          example: 2499.99
        shippingAddress:
          $ref: "#/components/schemas/Address"
          description: Optional shipping address (falls back to user's address)
...